Fix password validation error message, move CustomPasswordValidation into AuthCommon, add password error messages to Constants
