+++ OutOfSchool/OutOfSchool.AuthCommon/Resources/SharedResource.uk.resx
@@ -293,8 +293,8 @@
   <data name="Password confirmation is required" xml:space="preserve">
     <value>Потрібно підтвердження пароля</value>
   </data>
-  <data name="Password must contain at least one capital, number and symbol(@$!%*?&amp;)." xml:space="preserve">
-    <value>Помилка! Пароль має складатися мінімум з 8-ми символів, із них одна головна літера, цифри та спецсимволи (@$!%*? &amp;).</value>
+  <data name="Password must be at least 8 characters long, including upper and lower case letters, digits and special characters (@$!%*?&amp;)" xml:space="preserve">
+    <value>Пароль має містити мінімум 8 символів та обов'язково включати великі і малі літери, цифри, спецсимволи (@$!%*?&amp;)</value>
   </data>
   <data name="Check the entered data. Please use only cyrillic and symbols( ' - )" xml:space="preserve">
     <value>Перевірте введені дані. Використовуйте, будь ласка, тільки кирилицю та символи( ' - ).</value>
