+++ OutOfSchool/OutOfSchool.AuthCommon/ViewModels/ChangePasswordLoginViewModel.cs
@@ -14,7 +14,7 @@ public class ChangePasswordLoginViewModel : IValidatableObject
 
     [RegularExpression(
         Constants.PasswordRegexViewModel,
-        ErrorMessage = "Password must contain at least one capital, number and symbol(@$!%*?&).")]
+        ErrorMessage = Constants.PasswordValidationErrorMessage)]
     [DataType(DataType.Password)]
     public string NewPassword { get; set; }
